Moved ProjectDataStore to com.google.refine package
