+++ main/src/com/google/refine/ProjectManager.java
@@ -49,7 +49,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.refine.history.HistoryEntryManager;
-import com.google.refine.io.ProjectDataStore;
 import com.google.refine.model.Project;
 import com.google.refine.preference.PreferenceStore;
 import com.google.refine.preference.TopList;
