+++ main/src/com/google/refine/io/FileProjectDataStore.java
@@ -27,6 +27,7 @@ import org.json.JSONWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.refine.ProjectDataStore;
 import com.google.refine.ProjectMetadata;
 import com.google.refine.history.History;
 import com.google.refine.history.HistoryEntry;
